YUM and DNF global definitions
