+++ hotness/repository.py
@@ -1,61 +1,32 @@
 import logging
 import subprocess
 import os
-import ConfigParser
-from six import StringIO
-
 from hotness.cache import cache
 
 log = logging.getLogger('fedmsg')
 
-thn_section = 'thn'
 
-
-class ThnConfigParser(ConfigParser.ConfigParser):
-    def read(self, filename):
-        try:
-            text = open(filename).read()
-        except IOError:
-            pass
-        else:
-            section = "[%s]\n" % thn_section
-            file = StringIO(section + text)
-            self.readfp(file, filename)
-
-
-def get_pkg_manager():
-    release_file = '/etc/os-release'
-    config = ThnConfigParser()
-    config.read(release_file)
-    name = config.get(thn_section, 'ID')
-    if name == 'fedora':
-        return 'dnf'
-    else:
-        return 'yum'
-
-
-def get_version(package_name, yumconfig):
-    nvr_dict = build_nvr_dict(yumconfig)
+def get_version(package_name, yumconfig, pkg_manager):
+    nvr_dict = build_nvr_dict(yumconfig, pkg_manager)
     try:
         version = nvr_dict[package_name]
     except KeyError:
         log.warn("Did not find %r in nvr_dict, forcing refresh" % package_name)
-        force_cache_refresh(yumconfig)
-        nvr_dict = build_nvr_dict(yumconfig)
+        force_cache_refresh(yumconfig, pkg_manager)
+        nvr_dict = build_nvr_dict(yumconfig, pkg_manager)
         # This might still fail.. but we did the best we could.
         version = nvr_dict[package_name]
     return version
 
 
-def force_cache_refresh(yumconfig):
+def force_cache_refresh(yumconfig, pkg_manager):
     # First, invalidate our in-memory cache of the results
     cache.invalidate(hard=True)
 
     # But also ask yum/dnf to kill its on-disk cache
-    cmdline = [os.path.join("/usr/bin", get_pkg_manager()),
-               "--config", yumconfig,
-               "clean",
-               "all"]
+    cmdline = pkg_manager + ["--config", yumconfig,
+                             "clean",
+                             "all"]
     log.info("Running %r" % ' '.join(cmdline))
     cleanall = subprocess.Popen(cmdline, stdout=subprocess.PIPE)
     (stdout, stderr) = cleanall.communicate()
@@ -65,20 +36,13 @@ def force_cache_refresh(yumconfig):
 
 
 @cache.cache_on_arguments()
-def build_nvr_dict(yumconfig):
-    pkg_manager = get_pkg_manager()
-    cmdline = []
-    if pkg_manager == 'yum':
-        cmdline = ["/usr/bin/repoquery"]
-    else:
-        cmdline = [os.path.join("/usr/bin", pkg_manager),
-                   "repoquery"]
-    cmdline.extend(["--config", yumconfig,
-                    "--quiet",
-                    #"--archlist=src",
-                    "--all",
-                    "--qf",
-                    "%{name}\t%{version}\t%{release}"])
+def build_nvr_dict(yumconfig, pkg_manager):
+    cmdline = pkg_manager + ["--config", yumconfig,
+                             "--quiet",
+                             #"--archlist=src",
+                             "--all",
+                             "--qf",
+                             "%{name}\t%{version}\t%{release}"]
 
     log.info("Running %r" % ' '.join(cmdline))
     repoquery = subprocess.Popen(cmdline, stdout=subprocess.PIPE)
