+++ setup.py
@@ -14,7 +14,7 @@ setup(
         "dogpile.cache",
         "requests",
         "sh",
-        "python-six",
+        "six",
         "fedmsg_meta_fedora_infrastructure",
     ],
     packages=find_packages(),
